reset original jimm.go file
